+++ src/main/java/com/github/jlangch/venice/impl/util/excel/ExcelCellStyles.java
@@ -21,9 +21,12 @@
  */
 package com.github.jlangch.venice.impl.util.excel;
 
+import java.awt.Color;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.poi.hssf.usermodel.HSSFWorkbook;
+import org.apache.poi.hssf.util.HSSFColor;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.DataFormat;
 import org.apache.poi.ss.usermodel.FillPatternType;
@@ -31,6 +34,9 @@ import org.apache.poi.ss.usermodel.Font;
 import org.apache.poi.ss.usermodel.HorizontalAlignment;
 import org.apache.poi.ss.usermodel.VerticalAlignment;
 import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.usermodel.XSSFCellStyle;
+import org.apache.poi.xssf.usermodel.XSSFColor;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
 
 /**
@@ -91,7 +97,56 @@ public class ExcelCellStyles {
 
 		cellStyles.put(name, style);
 	}
-	
+
+	public void registerCellFormat(
+			final String name,
+			final String dataFormat,
+			final String fontRefName,
+			final Color bgColor,
+			final Boolean wrapText,
+			final HorizontalAlignment hAlign,
+			final VerticalAlignment vAlign
+	) {
+		if (name == null) {
+			throw new IllegalArgumentException("A cell format name must not be null");
+		}
+		
+		final CellStyle style = workbook.createCellStyle();
+		
+		if (bgColor != null) {
+			if (workbook instanceof XSSFWorkbook) {
+				((XSSFCellStyle)style).setFillForegroundColor(
+						new XSSFColor(bgColor, null));
+			}
+			else if (workbook instanceof HSSFWorkbook) {
+				final HSSFColor hssfColor = ColorUtil.bestHSSFColor((HSSFWorkbook)workbook, bgColor);
+				style.setFillForegroundColor(hssfColor.getIndex());
+			}
+			
+			style.setFillPattern(FillPatternType.SOLID_FOREGROUND);
+		}
+		if (dataFormat != null) {
+			style.setDataFormat(this.dataFormat.getFormat(dataFormat));
+		}
+		if (fontRefName != null) {
+			final Font font = fonts.get(fontRefName);
+			if (font != null) {
+				style.setFont(font);
+			}
+		}
+		if (wrapText != null) {
+			style.setWrapText(wrapText);
+		}
+		if (hAlign != null) {
+			style.setAlignment(hAlign);
+		}
+		if (vAlign != null) {
+			style.setVerticalAlignment(vAlign);
+		}
+
+		cellStyles.put(name, style);
+	}
+
 	public CellStyle getCellStyle(final String name) {
 		return (name != null) ? cellStyles.get(name) : null;
 	}
@@ -101,7 +156,7 @@ public class ExcelCellStyles {
 			.entrySet()
 			.stream()
 			.filter(e -> e.getValue() != null)
-			.forEach(e -> registerCellFormat(e.getKey(), e.getValue(), null, null, null, null, null));	
+			.forEach(e -> registerCellFormat(e.getKey(), e.getValue(), null, (Short)null, null, null, null));	
 	}
 
 	private Map<String,String> getStandardFormats() {